refactor: ai 요청 관련 내용 ai 도메인으로 분리
