+++ backend/src/main/java/aimo/backend/domains/ai/dto/parameter/JudgementToAiParameter.java
@@ -1,6 +1,5 @@
-package aimo.backend.domains.privatePost.dto.parameter;
+package aimo.backend.domains.ai.dto.parameter;
 
-import aimo.backend.domains.privatePost.dto.request.JudgementToAiRequest;
 import aimo.backend.domains.privatePost.model.OriginType;
 
 public record JudgementToAiParameter(
@@ -18,14 +17,6 @@ public record JudgementToAiParameter(
 			memberId,
 			content,
 			originType
-		);
-	}
-
-	public static JudgementToAiParameter from(Long memberId, JudgementToAiRequest request) {
-		return new JudgementToAiParameter(
-			memberId,
-			request.content(),
-			request.originType()
 		);
 	}
 }
