+++ backend/src/main/java/aimo/backend/domains/ai/dto/request/ImageToTextRequest.java
@@ -1,4 +1,4 @@
-package aimo.backend.domains.privatePost.dto.request;
+package aimo.backend.domains.ai.dto.request;
 
 import jakarta.validation.constraints.NotNull;
