+++ backend/src/main/java/aimo/backend/domains/ai/dto/response/ImageToTextResponse.java
@@ -1,4 +1,4 @@
-package aimo.backend.domains.privatePost.dto.response;
+package aimo.backend.domains.ai.dto.response;
 
 public record ImageToTextResponse(
 	String script
