+++ backend/src/main/java/aimo/backend/domains/ai/dto/response/SpeechToTextResponse.java
@@ -1,4 +1,4 @@
-package aimo.backend.domains.privatePost.dto.response;
+package aimo.backend.domains.ai.dto.response;
 
 public record SpeechToTextResponse(
 	String script
