+++ backend/src/main/java/aimo/backend/domains/privatePost/controller/PrivatePostController.java
@@ -8,8 +8,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,21 +18,9 @@ import aimo.backend.common.util.memberLoader.MemberLoader;
 import aimo.backend.domains.privatePost.dto.parameter.DeletePrivatePostParameter;
 import aimo.backend.domains.privatePost.dto.parameter.FindPrivatePostParameter;
 import aimo.backend.domains.privatePost.dto.parameter.FindPrivatePostPreviewParameter;
-import aimo.backend.domains.privatePost.dto.parameter.ImageToTextParameter;
-import aimo.backend.domains.privatePost.dto.parameter.JudgementToAiParameter;
-import aimo.backend.domains.privatePost.dto.parameter.SpeechToTextParameter;
-import aimo.backend.domains.privatePost.dto.parameter.UpdateContentToPrivatePostParameter;
-import aimo.backend.domains.privatePost.dto.request.ImageToTextRequest;
-import aimo.backend.domains.privatePost.dto.request.SpeechToTextRequest;
-import aimo.backend.domains.privatePost.dto.request.UpdateContentToPrivatePostRequest;
-import aimo.backend.domains.privatePost.dto.request.UploadTextRecordAndRequestJudgementRequest;
-import aimo.backend.domains.privatePost.dto.response.ImageToTextResponse;
 import aimo.backend.domains.privatePost.dto.response.PrivatePostPreviewResponse;
 import aimo.backend.domains.privatePost.dto.response.PrivatePostResponse;
-import aimo.backend.domains.privatePost.dto.response.SpeechToTextResponse;
-import aimo.backend.domains.privatePost.model.OriginType;
 import aimo.backend.domains.privatePost.service.PrivatePostService;
-import aimo.backend.domains.upload.service.FileService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -45,65 +31,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class PrivatePostController {
 
-	private final FileService fileService;
 	private final PrivatePostService privatePostService;
-
-	@PostMapping("/judgement/text")
-	public ResponseEntity<DataResponse<Void>> uploadTextRecordAndJudgement(
-		@Valid @RequestBody UploadTextRecordAndRequestJudgementRequest request
-	) {
-		Long memberId = MemberLoader.getMemberId();
-
-		JudgementToAiParameter judgementToAiParameter = JudgementToAiParameter.of(
-			memberId,
-			request.content(),
-			OriginType.TEXT);
-
-		privatePostService.uploadTextRecordAndRequestJudgement(judgementToAiParameter);
-
-		return ResponseEntity.ok(DataResponse.ok());
-	}
-
-	// AI로부터의 판결문 요청 콜백
-	@PostMapping("/judgement/callback")
-	public ResponseEntity<DataResponse<Void>> updateContentToPrivatePost(
-		@Valid @RequestBody UpdateContentToPrivatePostRequest request
-	) {
-		UpdateContentToPrivatePostParameter parameter = UpdateContentToPrivatePostParameter.from(request);
-		privatePostService.updateContentToPrivatePost(parameter);
-
-		return ResponseEntity.ok(DataResponse.ok());
-	}
-
-	// @PostMapping("/chat")
-	// public ResponseEntity<DataResponse<SavePrivatePostResponse>> uploadChatRecord(
-	// 	@Valid @RequestParam("chat_record") ChatRecordRequest chatRecordRequest
-	// ) throws IOException {
-	//
-	//
-	// 	return ResponseEntity.status(HttpStatus.CREATED)
-	// 		.body(DataResponse.created(new SavePrivatePostResponse(privatePostId)));
-	// }
-
-	@PostMapping("/speech-to-text")
-	public ResponseEntity<DataResponse<SpeechToTextResponse>> speechToText(
-		@Valid @RequestBody SpeechToTextRequest request
-	) {
-		SpeechToTextParameter parameter = SpeechToTextParameter.from(request);
-
-		return ResponseEntity.status(HttpStatus.CREATED)
-			.body(DataResponse.created(fileService.speechToText(parameter)));
-	}
-
-	@PostMapping("/image-to-text")
-	public ResponseEntity<DataResponse<ImageToTextResponse>> ImageToText(
-		@Valid @RequestBody ImageToTextRequest request
-	) {
-		ImageToTextParameter parameter = ImageToTextParameter.from(request.url());
-
-		return ResponseEntity.status(HttpStatus.CREATED)
-			.body(DataResponse.created(fileService.imageToText(parameter)));
-	}
 
 	// 개인글 조회
 	@GetMapping("/{privatePostId}")
