+++ backend/src/main/java/aimo/backend/domains/privatePost/dto/request/UpdateContentToPrivatePostRequest.java
@@ -1,6 +1,5 @@
 package aimo.backend.domains.privatePost.dto.request;
 
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record UpdateContentToPrivatePostRequest(
