--- backend/src/main/java/aimo/backend/domains/privatePost/dto/response/SavePrivatePostResponse.java
@@ -1,8 +1,0 @@
-package aimo.backend.domains.privatePost.dto.response;
-
-public record SavePrivatePostResponse(Long privatePostId) {
-
-	public static SavePrivatePostResponse of(Long privatePostId) {
-		return new SavePrivatePostResponse(privatePostId);
-	}
-}
