+++ backend/src/main/java/aimo/backend/domains/upload/service/FileService.java
@@ -9,13 +9,9 @@ import aimo.backend.common.properties.AiServerProperties;
 import aimo.backend.common.util.webclient.ReactiveHttpService;
 import aimo.backend.domains.member.entity.Member;
 import aimo.backend.domains.member.repository.MemberRepository;
-import aimo.backend.domains.privatePost.dto.parameter.ImageToTextParameter;
-import aimo.backend.domains.privatePost.dto.response.ImageToTextResponse;
-import aimo.backend.domains.privatePost.dto.response.SpeechToTextResponse;
 import aimo.backend.domains.upload.dto.parameter.CreateFilePreSignedUrlParameter;
 import aimo.backend.domains.upload.dto.parameter.SaveFileMetaDataParameter;
 import aimo.backend.domains.upload.dto.parameter.SaveProfileImageMetaDataParameter;
-import aimo.backend.domains.privatePost.dto.parameter.SpeechToTextParameter;
 import aimo.backend.domains.upload.dto.response.FindProfileImageUrlResponse;
 import aimo.backend.domains.upload.dto.response.SaveFileMetaDataResponse;
 import aimo.backend.domains.upload.entity.FileRecord;
@@ -32,25 +28,11 @@ import lombok.RequiredArgsConstructor;
 public class FileService {
 
 	private final FileRepository fileRepository;
-	private final ReactiveHttpService reactiveHttpService;
-	private final AiServerProperties aiServerProperties;
 	private final S3Service s3Service;
 	private final MemberRepository memberRepository;
 	private final ProfileImageRepository profileImageRepository;
 
-	// AI서버에 음성 파일을 텍스트로 변환 요청
-	public SpeechToTextResponse speechToText(SpeechToTextParameter speechToTextParameter) {
-		String url = aiServerProperties.getDomainUrl() + aiServerProperties.getSpeechToTextApi();
-		return reactiveHttpService.post(url, speechToTextParameter, SpeechToTextResponse.class).block();
-	}
-
-	// AI서버에 이미지 파일을 텍스트로 변환 요청
-	public ImageToTextResponse imageToText(ImageToTextParameter parameter) {
-		String url = aiServerProperties.getDomainUrl() + aiServerProperties.getImageToTextApi();
-		return reactiveHttpService.post(url, parameter, ImageToTextResponse.class).block();
-	}
-
-	// 음성 파일 메타데이터 저장
+	// 파일 메타데이터 저장
 	@Transactional(rollbackFor = ApiException.class)
 	public SaveFileMetaDataResponse saveFileMetaData(SaveFileMetaDataParameter parameter) {
 		// url 생성
